Fetch Pokemon species details
Create a components directory
and a PokemonList file
inside it.

In the file, import useState
and useEffect. Create a
PokemonList component
that fetches the
species data.

Create a pokemonLimit
constant that would
be passed into the
url to set the
fetch results limit.

In a useEffect, define
the fetchPokemon
function. Pass
the pokemonLimit
into the url.
Fetch the speciesResponse
in a try catch block
and throw an error
if the response is not
ok.

Await the parsed data
with speciesList.
The data (object)
contains a
results array that
contains the name
and url of
the requested pokemons
in different objects.

Loop through this
array each object's url
with fetchPromises.

Await all the promises
with pokemonResponses.
Promise.all returns
an array of responses..

Loop through the array
and throw an error if
any response is not ok.

Loop through
pokemonResponses and call
the .json() async function
on each of the responses
in jsonPromises.

Await all the jsonPromises
in pokemonDetails and set the
pokemonList state with it.

Call fetchPokemon() and
add pokemonLimit to the
dependency array.