Display Pokemon details and image
Create a
getPokemonImageUrl that
prepares the pokemon image
url. It takes in an id
and pads the id with two
leading zeros as the id
must be zero-padded, three
digit pokemon.It passes
this padded it into the
url and returns it.
The function just makes
sure the id is 3 digits
long and pads the right
number of 0s if it is
not.

This ID corresponds to the
ID as fetched from the
API is the previous
commit.

In the return statement,
loop through the
pokemonList and
for each pokemon,
create a div.
Inside the div, display
the pokemon name
and create a image
that its source
is the url generated
from getPokemonImageUrl
with the respective
pokemon id passed in.

Include PokemonList in
main.jsx.