Shuffle state array on click
Use the Fisher Yates
algorithm in the
shuffle(array) function.
Create a copy of
the array passed in
so the function
will not mutata state
and so that react
can see a new object
in memory and then
rerender.

In a handleClick
handler, use the
functional state
update to
call the shuffle
function on
the state and
then set the
state to the
returned array.

In the return
statement, set
the handleClick
to the div's
onClick event.
The state array is
shuffled whenever
any of the pokemons
is clicked on and
the display is
shuffled accordingly.