Add props to PokemonList
pokemonLimit would be
used to set the limit
of results.

onPokemonSelect is a
callback that would
be passed down from
the parent to inform
the parent about the
id of the pokemon
that was selected(clicked),

currentScore and highScore
will hold the current
and high scores
respectively.

In handleClick, call
onPokemonSelect with
the target's id.

In the return statement,
add a scoreboard
div with current score
and high score.