Implement the PokemonGame
In components directory,
create a PokemonGame file
and in it, import the
PokemonList.

Create a PokemonGame
component that
has 2 states.

selectedPokemonIds
is an array that holds
the ids of all pokemons
that have been clicked
on or selected in a
particular round.

score is an object
that hold the current
and high scores.

In a handleGuess function,
check if the currentId
is in the
selectedPokemonIds array.

If it is not, add it to
the array and increase
the current score.
If it is (means the pokemon
has once been clicked on
in this round), empty the
selectedPokemonIds array,
reset the currentScore
and set the highScore to
be the larger value between
the highScore and the
currentScore.
This ends one round.

In the return statement,
call the PokemonList
component and pass in
the props as required.