Add more styles
First, in PokemonList.jsx
change the current and
high score divs to
paragraphs and update
their classes.

In the css file, in
the game container,
increase the bottom
padding, add a
background color and
a color.

Add font-family
to title-and-scoreboard
game-description and
pokemon-name.

increase the font-size
of scoreboard,
game-description
and pokemon-name.

Add a gap to
pokemon-container.