Make title and scoreboard sticky
Remove the padding-top
from game container
and add a padding
instead to
title-and-scoreboard.
This is so that
title-and-scoreboard
can stay at the
top of the page when
sticky.

Add a background
color to
title-and-scoreboard
so that it is
not transparent
when text scrolls
behind it.
Add a top value,
and a sticky
position.
Add a z-index so
that every pokemon
that scales on
hover, stays
behind it.